+++ packages/cloud_firestore/cloud_firestore_web/lib/src/interop/firestore_interop.dart
@@ -7,6 +7,7 @@
 @JS('firebase.firestore')
 library firebase_interop.firestore;
 
+import './firestore.dart';
 import 'package:firebase_core_web/firebase_core_web_interop.dart';
 import 'dart:typed_data' show Uint8List;
 
@@ -42,6 +43,7 @@ abstract class FirestoreJsImpl {
   external PromiseJsImpl<Null> enablePersistence([PersistenceSettings? settings]);
 
   external void Function() onSnapshotsInSync(dynamic observer);
+
 // ignore: prefer_void_to_null
   external PromiseJsImpl<Null> clearPersistence();
 
@@ -61,7 +63,9 @@ abstract class FirestoreJsImpl {
 // ignore: prefer_void_to_null
   external PromiseJsImpl<Null> waitForPendingWrites();
 
-  external PromiseJsImpl<Null> setLogLevel(String logLevel);
+  external LoadBundleTaskJsImpl loadBundle(Uint8List bundle);
+
+  external PromiseJsImpl<QueryJsImpl?> namedQuery(String name);
 }
 
 @JS('WriteBatch')
@@ -223,6 +227,7 @@ abstract class DocumentReferenceJsImpl {
   external set path(String v);
 
   external CollectionReferenceJsImpl collection(String collectionPath);
+
 //ignore: prefer_void_to_null
   external PromiseJsImpl<Null> delete();
 
@@ -239,6 +244,32 @@ abstract class DocumentReferenceJsImpl {
 
 //ignore: prefer_void_to_null
   external PromiseJsImpl<Null> update(dynamic dataOrFieldsAndValues);
+}
+
+@JS('LoadBundleTask')
+abstract class LoadBundleTaskJsImpl {
+  external void Function() onProgress(
+    void Function(LoadBundleTaskProgressJsImpl) progress,
+  );
+
+  external PromiseJsImpl then([
+    Func1? onResolve,
+    dynamic Function(FirestoreError) onReject,
+  ]);
+}
+
+@JS()
+@anonymous
+abstract class LoadBundleTaskProgressJsImpl {
+  external String get bytesLoaded;
+
+  external int get documentsLoaded;
+
+  external String get taskState;
+
+  external String get totalBytes;
+
+  external int get totalDocuments;
 }
 
 @JS('DocumentSnapshot')
@@ -306,7 +337,7 @@ abstract class QueryJsImpl {
       /*DocumentSnapshot|List<dynamic>*/
       dynamic snapshotOrFieldValues);
 
-  external PromiseJsImpl<QuerySnapshotJsImpl> get();
+  external PromiseJsImpl<QuerySnapshotJsImpl> get([GetOptions? options]);
 
   external QueryJsImpl limit(num? limit);
 
@@ -535,6 +566,7 @@ abstract class SetOptions {
   external bool get merge;
 
   external set merge(bool v);
+
 //ignore: avoid_setters_without_getters
   external set mergeFields(List<String> v);