+++ packages/firebase_auth/firebase_auth/example/test_driver/instance_e2e.dart
@@ -122,6 +122,7 @@ void runInstanceTests() {
       tearDown(() async {
         await subscription.cancel();
       });
+
       test('calls callback with the current user and when user state changes',
           () async {
         await ensureSignedIn(testEmail);
@@ -141,11 +142,14 @@ void runInstanceTests() {
         }, count: 2, reason: 'Stream should only have been called 2 times'));
 
         await FirebaseAuth.instance.currentUser
-            .updateProfile(displayName: 'updatedName');
+            .updateDisplayName('updatedName');
 
         await FirebaseAuth.instance.currentUser.reload();
-        expect(FirebaseAuth.instance.currentUser.displayName,
-            equals('updatedName'));
+
+        expect(
+          FirebaseAuth.instance.currentUser.displayName,
+          equals('updatedName'),
+        );
       });
     });
 
@@ -688,5 +692,19 @@ void runInstanceTests() {
         expect(credential, isA<PhoneAuthCredential>());
       }, skip: kIsWeb || defaultTargetPlatform != TargetPlatform.android);
     }, skip: defaultTargetPlatform == TargetPlatform.macOS || kIsWeb);
+
+    group('tenantId', () {
+      test('User associated with the tenantId correctly', () async {
+        // tenantId created in the GCP console
+        const String tenantId = 'auth-tenant-test-xukxg';
+        // created User on GCP console associated with the above tenantId
+        final userCredential = await FirebaseAuth.instance
+            .signInWithEmailAndPassword(
+                email: 'test-tenant@email.com', password: 'fake-password');
+
+        expect(userCredential.user.tenantId, tenantId);
+      });
+      // todo(russellwheatley85): get/set tenantId and authenticating user via auth emulator is not possible at the moment.
+    }, skip: true);
   });
 }