+++ packages/firebase_messaging/firebase_messaging_platform_interface/lib/src/method_channel/method_channel_messaging.dart
@@ -128,6 +128,12 @@ class MethodChannelFirebaseMessaging extends FirebaseMessagingPlatform {
     return _autoInitEnabled;
   }
 
+  /// Returns "true" as this API is used to inform users of web browser support
+  @override
+  bool isSupported() {
+    return true;
+  }
+
   @override
   Future<RemoteMessage?> getInitialMessage() async {
     try {
@@ -164,11 +170,10 @@ class MethodChannelFirebaseMessaging extends FirebaseMessagingPlatform {
   }
 
   @override
-  Future<void> deleteToken({
-    String? senderId,
-  }) async {
+  Future<void> deleteToken() async {
     try {
-      await channel.invokeMapMethod('Messaging#deleteToken', {'appName': app.name, 'senderId': senderId});
+      await channel
+          .invokeMapMethod('Messaging#deleteToken', {'appName': app.name});
     } catch (e) {
       throw convertPlatformException(e);
     }
@@ -193,13 +198,12 @@ class MethodChannelFirebaseMessaging extends FirebaseMessagingPlatform {
 
   @override
   Future<String?> getToken({
-    String? senderId,
     String? vapidKey, // not used yet; web only property
   }) async {
     try {
-      Map<String, String?>? data = await channel.invokeMapMethod<String, String>('Messaging#getToken', {
+      Map<String, String?>? data =
+          await channel.invokeMapMethod<String, String>('Messaging#getToken', {
         'appName': app.name,
-        'senderId': senderId,
       });
 
       return data?['token'];