+++ packages/cloud_functions/cloud_functions_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 5.0.7
+
+ - Update a dependency to the latest release.
+
+## 5.0.6
+
+ - Update a dependency to the latest release.
+
+## 5.0.5
+
+ - Update a dependency to the latest release.
+
 ## 5.0.4
 
  - **DOCS**: Add missing homepage/repository links (#6054).