+++ packages/firebase_in_app_messaging/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.5.0+6
+
+ - Update a dependency to the latest release.
+
+## 0.5.0+5
+
+ - Update a dependency to the latest release.
+
+## 0.5.0+4
+
+ - Update a dependency to the latest release.
+
 ## 0.5.0+3
 
  - **DOCS**: Add missing homepage/repository links (#6054).